@@ -10,29 +10,25 @@
     "stylelint": "stylelint 'app/javascript/**/*.css' --fix",
     "test": "jest"
   },
-  "babel": {
-    "presets": [
-      [
-        "@babel/preset-env"
-      ]
-    ]
-  },
-  "browserslist": [
-    "defaults"
-  ],
   "dependencies": {
     "@babel/cli": "^7.4.4",
     "@babel/core": "^7.4.4",
+    "@babel/plugin-transform-runtime": "^7.4.4",
+    "@babel/polyfill": "^7.4.4",
     "@babel/preset-env": "^7.4.4",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/runtime": "^7.4.4",
     "@ledgerhq/hw-app-btc": "^4.33.5",
     "@ledgerhq/hw-transport-u2f": "^4.32.0",
-    "@rails/webpacker": "^3.5.5",
+    "@rails/webpacker": "^4.0.2",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "babel-preset-react": "^6.24.1",
     "bip39": "^2.5.0",
     "borc": "^2.0.4",
     "cardano-crypto.js": "^5.0.0",
     "classnames": "^2.2.6",
     "coffeescript": "1.12.7",
+    "core-js": "2",
     "d3": "3",
     "d3pie": "^0.1.9",
     "decimal.js": "^10.1.1",
@@ -40,15 +36,16 @@
     "formdata-polyfill": "^3.0.13",
     "jquery": "^3.3.1",
     "js-cookie": "^2.2.0",
-    "prop-types": "^15.6.2",
+    "postcss-preset-env": "^6.6.0",
+    "prop-types": "^15.7.2",
     "qtum-opcodes": "^1.0.0",
     "qtumjs-lib": "^0.0.7",
     "qweb3": "^1.2.2",
     "rails-ujs": "^5.2.2",
-    "react": "^16.6.0",
+    "react": "^16.8.6",
     "react-app-polyfill": "^1.0.0",
     "react-beautiful-dnd": "^10.0.3",
-    "react-dom": "^16.6.0",
+    "react-dom": "^16.8.6",
     "react-paginate": "^5.2.4",
     "react-pluralize": "^1.5.0",
     "react-slick": "^0.23.2",
@@ -62,13 +59,16 @@
     "styled-components": "^4.2.0",
     "tezos-wallet": "^1.1.3",
     "trezor-connect": "^6.0.3",
+    "uglifyjs-webpack-plugin": "^2.1.2",
     "wallet-address-validator": "^0.2.4",
     "whatwg-fetch": "^3.0.0"
   },
   "devDependencies": {
+    "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^7.2.1",
-    "babel-jest": "^23.6.0",
+    "babel-jest": "^24.7.1",
     "babel-plugin-rewire": "^1.2.0",
+    "babel-plugin-styled-components": "^1.10.0",
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
     "enzyme": "^3.7.0",
@@ -85,14 +85,17 @@
     "eslint-plugin-react-hooks": "^1.1.0-alpha.1",
     "eslint-plugin-standard": "^4.0.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^23.6.0",
+    "jest": "^24.7.1",
     "jest-enzyme": "^7.0.1",
     "jest-fetch-mock": "^2.1.0",
+    "postcss-cssnext": "^3.1.0",
+    "postcss-import": "^12.0.1",
+    "postcss-loader": "^3.0.0",
     "react-test-renderer": "^16.6.0",
-    "regenerator-runtime": "^0.12.1",
+    "regenerator-runtime": "^0.13.2",
     "stylelint": "^9.9.0",
     "stylelint-config-recommended": "^2.1.0",
-    "webpack-dev-server": "^2.9.4"
+    "webpack-dev-server": "^3.3.1"
   },
   "stylelint": {
     "extends": "stylelint-config-recommended"
@@ -116,7 +119,16 @@
       ".+\\.(css|styl|less|sass|scss)$": "identity-obj-proxy",
       ".+\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/spec/javascript/mockAssets.js"
     },
-    "setupTestFrameworkScriptFile": "./spec/javascript/setupFramework.js",
+    "moduleFileExtensions": [
+      "js",
+      "jsx"
+    ],
+    "transform": {
+      "^.+\\.(js|jsx)$": "./spec/javascript/transform.js"
+    },
+    "setupFilesAfterEnv": [
+      "./spec/javascript/setupFramework.js"
+    ],
     "automock": false,
     "setupFiles": [
       "./spec/javascript/setupJest.js"
