@@ -12,9 +12,9 @@ export default {
       let tx = unspentTransactions[i]
       findTotal = findTotal.plus(tx.satoshis)
       find[find.length] = tx
-      if (findTotal.isGreaterThanOrEqualTo(value)) break
+      if (findTotal.gte(value)) break
     }
-    if (value.isGreaterThan(findTotal)) {
+    if (value.gt(findTotal)) {
       throw new Error('You do not have enough tokens to send')
     }
     return find
